requirements.txt
nextcord>=2.5.0
python-dotenv